+++ mods/aqualumen-ui/src/main/java/store/aquateche/aqualumen/AquaLumenUI.java
@@ -27,7 +27,7 @@ import store.aquateche.aqualumen.registry.ModRegistries;
 public final class AquaLumenUI {
 
     public static final String MODID = "aqualumen";
-    public static final String VERSION = "0.3.37-alpha";
+    public static final String VERSION = "0.3.46-alpha";
     public static final Logger LOGGER = LogUtils.getLogger();
 
     public AquaLumenUI() {